+++ net.resheim.eclipse.timekeeper.db/src/main/java/net/resheim/eclipse/timekeeper/db/adapter/EmbeddedTimekeeperService.java
@@ -11,6 +11,6 @@ public final class EmbeddedTimekeeperService extends DefaultTimekeeperService {
 	}
 
 	EmbeddedTimekeeperService(JpaServicePorts ports) {
-		super(ports.ports(event -> TimekeeperPlugin.getDefault().notifyServiceListeners()));
+		super(ports.ports(event -> ports.afterCommit(() -> TimekeeperPlugin.getDefault().notifyServiceListeners())));
 	}
 }